Added new crop, new trees, and more
I have added the sphenophyllum crop, but it can only be planted on
farmland due to the way BlockCrops is written. I have also added new
logs etc for trees, but the rotten logs have a missing texture because I
wasn't given a texture for the rotten logs.

I  have also done a bit more code cleanup and fixing, and removed the
planks, slabs, etc. with the current trees.